General restructuring, refactoring, new branch
