+++ wwu_wash_and_dry/lib/class/floor_layout.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'home_widgets/machine.dart';
+import 'machine.dart';
 // Created by Reece Reklai and Bradon Ladd
 
 // Test data to get a mock up and running
