+++ wwu_wash_and_dry/lib/class/machine.dart
