+++ wwu_wash_and_dry/test/login_test.dart
@@ -7,12 +7,12 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:wwu_wash_and_dry/home.dart';
+import 'package:wwu_wash_and_dry/lib/main.dart';
 
 // import 'package:wwu_wash_and_dry/home_widgets/home_page.dart';
 // import 'package:wwu_wash_and_dry/home_widgets/login_error.dart';
 // import 'package:wwu_wash_and_dry/home_widgets/login_page.dart';
-import 'package:wwu_wash_and_dry/home_widgets/login.dart';
+import 'package:wwu_wash_and_dry/widgets/login.dart';
 
 void main() {
   testWidgets('Login returns a Login_page on recievieng a \'-1\'',
